Towards restructuring code 1) case-specific b.c. and radiation all included within drivers 2) cases touched: rayleigh-benard and dycoms 3) dycoms moved from examples/DGmethods to examples/Atmos
