+++ src/Atmos/Model/AtmosModel.jl
@@ -72,8 +72,6 @@ function vars_diffusive(m::AtmosModel, FT)
     moisture::vars_diffusive(m.moisture,FT)
   end
 end
-
-
 function vars_aux(m::AtmosModel, FT)
   @vars begin
     ∫dz::vars_integrals(m, FT)
@@ -91,16 +89,6 @@ function vars_integrals(m::AtmosModel,FT)
     radiation::vars_integrals(m.radiation,FT)
   end
 end
-
-include("orientation.jl")
-include("ref_state.jl")
-include("turbulence.jl")
-include("moisture.jl")
-include("radiation.jl")
-include("source.jl")
-include("boundaryconditions.jl")
-include("linear.jl")
-include("remainder.jl")
 
 """
     flux_nondiffusive!(m::AtmosModel, flux::Grad, state::Vars, aux::Vars,
@@ -132,12 +120,7 @@ Where
 
   # pressure terms
   p = pressure(m.moisture, m.orientation, state, aux)
-
-  if m.ref_state isa HydrostaticState
-    flux.ρu += (p-aux.ref_state.p)*I
-  else
-    flux.ρu += p*I
-  end
+  flux.ρu += p*I
   flux.ρe += u*p
   flux_radiation!(m.radiation, flux, state, aux, t)
   flux_moisture!(m.moisture, flux, state, aux, t)
@@ -222,6 +205,16 @@ end
 function integrate_aux!(m::AtmosModel, integ::Vars, state::Vars, aux::Vars)
   integrate_aux!(m.radiation, integ, state, aux)
 end
+
+include("orientation.jl")
+include("ref_state.jl")
+include("turbulence.jl")
+include("moisture.jl")
+include("radiation.jl")
+include("source.jl")
+include("boundaryconditions.jl")
+include("linear.jl")
+include("remainder.jl")
 
 # TODO: figure out a nice way to handle this
 function init_aux!(m::AtmosModel, aux::Vars, geom::LocalGeometry)
